Merge 'feat/shoppingCart' into 'main'

+++ app/Http/Controllers/BuyController.php
@@ -2,10 +2,13 @@
 
 namespace App\Http\Controllers;
 
+use App\Exceptions\LackofProductAmountException;
 use App\Models\BuyHistory;
 use App\Models\BuyRecord;
 use App\Models\CartItem;
 use App\Models\Product;
+use http\Message;
+use Exception;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
@@ -44,7 +47,9 @@ class BuyController extends Controller
                     $buyItems = Auth::user()->cartItems;
                 else
                     $buyItems = Auth::user()->cartItems()->whereIn('id', $itemIDs)->get();
-
+                if($buyItems->count() <= 0){
+                    throw new Exception("Transaction fail");
+                }
                 // Create new buy history
                 $buyHistory->save();
 
@@ -55,12 +60,23 @@ class BuyController extends Controller
                         'product_id' => $item->product->id,
                         'quantity' => $item->quantity,
                     ]);
-
+                    $product = Product::where('id', $item->product->id)->first();
+                    // Check if the product amount is smaller than the item quantity
+                    if ($product->amount < $item->quantity) {
+                        throw new LackofProductAmountException();
+                    }
+                    $product->update(['amount'=>$product->amount-$item->quantity]);
                     $item->delete();
                 }
             });
+        } catch(LackofProductAmountException $e){
+            $e->report();
+            return redirect()->back()->with('error', '含有庫存不足的商品！');
         } catch (\Throwable $e) {
-            return response()->noContent()->setStatusCode(400);
+            $cartItems = Auth::user()->cartItems()->orderBy('created_at')->get();
+            return view('payment', [
+                'cartItems' => $cartItems,
+            ]);
         }
         return view('bill', [
             'buyHistory' => $buyHistory,
